refactor: Reorganize mock component and rename for clarity
Implement a directory specific to mocks to enhance readability
and organization of test suites.

Rename MockStateEffect to MockLayoutAppLazyEffect for more
descriptive naming.
