+++ components/auth/authEffect/__test__/__mock__/mockErrorMessageObserver.tsx
@@ -1,0 +1,13 @@
+import { atomAuthErrorMessage } from '@auth/auth.states';
+import { useEffect } from 'react';
+import { useRecoilState } from 'recoil';
+
+export const MockErrorMessageObserver = () => {
+  const [message, setMessage] = useRecoilState(atomAuthErrorMessage);
+  const defaultError = 'Something went wrong';
+
+  useEffect(() => {
+    setMessage(defaultError);
+  }, [setMessage]);
+  return <>{message}</>;
+};
